Fix num-to-double type errors in clampRect and drop generated test stub
Dart's num.clamp returns num, which cannot be passed to
Rect.fromLTRB's double parameters. Also remove the CI-generated
widget_test.dart that references a nonexistent MyApp class.

+++ dermaview/app/lib/main.dart
@@ -656,10 +656,10 @@ class _ScannerScreenState extends State<ScannerScreen>
 
   ui.Rect clampRect(ui.Rect rect, int width, int height) {
     return ui.Rect.fromLTRB(
-      rect.left.clamp(0, width - 1),
-      rect.top.clamp(0, height - 1),
-      rect.right.clamp(1, width),
-      rect.bottom.clamp(1, height),
+      rect.left.clamp(0, width - 1).toDouble(),
+      rect.top.clamp(0, height - 1).toDouble(),
+      rect.right.clamp(1, width).toDouble(),
+      rect.bottom.clamp(1, height).toDouble(),
     );
   }
